fix: Move test to nimbus
